Import from 4.2.21. Add new makefile structure with configs dir.
git-svn-id: https://svn.ufz.de/svn/ogs/trunk/sources@28 3895a583-e809-0410-9bb6-aa9463026377

+++ FEM/rf_react.cpp
@@ -231,7 +231,7 @@ for(i=0;i<vector_size;i++){
 	m_pcs = pcs_vector[i];
 	if(m_pcs->pcs_type_name.compare("MASS_TRANSPORT")==0) np++;
 	if(m_pcs->pcs_type_name.compare("HEAT_TRANSPORT")==0) heatflag = 1;
-  nodenumber = m_pcs->m_msh->GetNodesNumber(false);
+	nodenumber = (long) m_pcs->m_msh->GetNodesNumber(false);
  }
  number_of_comp = np;
